Move resizeDerivativesToMatchScalar to autodiff.h
